+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based now on [Keep a
 Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.1.0]
+### Breaking changes
+- `Uploadcare\Interfaces\UploaderInterface::checkStatus(string $token)` now returns `Uploadcare\Interfaces\File\FileInfoInterface`
+### Added
+- `Uploadcare\Interfaces\UploaderInterface::syncUploadFromUrl` method for synchronically uploading a file from a remote URL
+- `Uploadcare\Interfaces\UploaderInterface::getSignature` method for get signature to use in third-party uploaders
+### Removed
+- `Uploadcare\Interfaces\Api\GroupApiInterface::storeGroup($id)`
+- `Uploadcare\Group::store()`
+
 ## [4.0.2]
 ### Added
 - Method `removeKey` to remove medatada keys